+++ src/nl/rug/ai/mas/oops/formula/Agent.java
@@ -19,11 +19,9 @@
 
 package nl.rug.ai.mas.oops.formula;
 
-import java.math.BigInteger;
-
 public interface Agent {
 	public Substitution<Agent> match(Agent other);
 	public Agent substitute(Substitution<Agent> s);
 	public boolean isConcrete();
-	public BigInteger code();
+	public int code();
 }
